refactor: extract Participation.Attendance from the context facade
Second of four. Attendance owns ParticipationRecord and AttendanceTransition:
roster seeding, check-in/out, manual absence, correction, and the history reads.
The facade drops from 1277 to 642 lines; 482 participation tests and 1167 web
tests pass with no test file edited.

Four functions are public because something outside this module needs them, and
each says so in its own doc rather than leaving the reader to guess:

  authorize_for_record/2         the single ADR-0017 guard. SessionNotes asks the
                                 same question -- a note about a child at a
                                 session is authorized by that session -- so it
                                 calls this rather than growing a second copy.
  mark_roster_absent_for_session/1  the roster half of completing a session. It
                                 opens no transaction: the caller supplies one,
                                 because the session row is the other half of the
                                 same write.
  list_records_by_session/1      for the roster reads still in the facade.
  absence_reasons_for_records/1  same, for why a child was marked absent.

normalize_notes/1 is duplicated here as a private function instead of being
shared. It is six lines of pure trimming with no invariant that can drift, and a
module existing only to hold it would be exactly the helpers/ catch-all the
layout rules reject. It also stops being public on the facade, where it was
@doc false with no caller in lib or test.
